Finally fixes the annoying bug that prevented chaos to work on machines with 16 meg or less RAM. See the diff for the fix -- the problem was with pointer arithmetic (not the first time). NEVER DO THIS AGAIN! Always cast pointers to address_t before doing arithmetic with it. Okay?
git-svn-id: https://svn.code.sf.net/p/chaos/svn/trunk/chaos@530 1f819a86-672b-0410-b5ed-9de735de380c

+++ stormG3/source/ia32/init.c
@@ -1,4 +1,4 @@
-/* $chaos: init.c,v 1.8 2002/08/07 18:21:17 per Exp $ */
+/* $chaos: init.c,v 1.9 2002/08/07 18:24:10 per Exp $ */
 /* Abstract: storm initialization. */
 /* Author: Per Lundberg <per@chaosdev.org> 
            Henrik Hallin <hal@chaosdev.org> */
@@ -167,7 +167,7 @@ void _start (void)
          :
          :
          "g" (KERNEL_DATA_SELECTOR),
-         "g" (&kernel_stack + KERNEL_STACK_SIZE));
+         "g" ((uint32_t) &kernel_stack + KERNEL_STACK_SIZE));
 
     /* Pass control to the kernel. */
     asm ("ljmp   %0, %1"
